refactor: move Measurement constructor inside model
